Move search clause types to their own file
